requirements.txt
flask
gunicorn
tensorflow
numpy
pillow
python-dotenv
groq
matplotlib
reportlab
